fix: single cacert symbol — route GH downloads through repository.c
